+++ crates/rust-client/src/rpc/mod.rs
@@ -56,7 +56,7 @@ use miden_objects::address::NetworkId;
 use miden_objects::block::{BlockHeader, BlockNumber, ProvenBlock};
 use miden_objects::crypto::merkle::{MmrProof, SmtProof};
 use miden_objects::note::{NoteId, NoteScript, NoteTag, Nullifier};
-use miden_objects::transaction::ProvenTransaction;
+use miden_objects::transaction::{ProvenTransaction, TransactionInputs};
 
 /// Contains domain types related to RPC requests and responses, as well as utility functions
 /// for dealing with them.
@@ -78,6 +78,9 @@ mod tonic_client;
 #[cfg(feature = "tonic")]
 pub use tonic_client::GrpcClient;
 
+use crate::rpc::domain::account_vault::AccountVaultInfo;
+use crate::rpc::domain::storage_map::StorageMapInfo;
+use crate::rpc::domain::transaction::TransactionsInfo;
 use crate::store::InputNoteRecord;
 use crate::store::input_note_states::UnverifiedNoteState;
 use crate::transaction::ForeignAccount;
@@ -103,6 +106,7 @@ pub trait NodeRpcClient: Send + Sync {
     async fn submit_proven_transaction(
         &self,
         proven_transaction: ProvenTransaction,
+        transaction_inputs: TransactionInputs,
     ) -> Result<BlockNumber, RpcError>;
 
     /// Given a block number, fetches the block header corresponding to that height from the node
@@ -301,6 +305,48 @@ pub trait NodeRpcClient: Send + Sync {
     /// - [`RpcError::ExpectedDataMissing`] if the note with the specified root is not found.
     async fn get_note_script_by_root(&self, root: Word) -> Result<NoteScript, RpcError>;
 
+    /// Fetches storage map updates for specified account and storage slots within a block range,
+    /// using the `/SyncStorageMaps` RPC endpoint.
+    ///
+    /// - `block_from`: The starting block number for the range.
+    /// - `block_to`: The ending block number for the range.
+    /// - `account_id`: The account ID for which to fetch storage map updates.
+    async fn sync_storage_maps(
+        &self,
+        block_from: BlockNumber,
+        block_to: Option<BlockNumber>,
+        account_id: AccountId,
+    ) -> Result<StorageMapInfo, RpcError>;
+
+    /// Fetches account vault updates for specified account within a block range,
+    /// using the `/SyncAccountVault` RPC endpoint.
+    ///
+    /// - `block_from`: The starting block number for the range.
+    /// - `block_to`: The ending block number for the range.
+    /// - `account_id`: The account ID for which to fetch storage map updates.
+    async fn sync_account_vault(
+        &self,
+        block_from: BlockNumber,
+        block_to: Option<BlockNumber>,
+        account_id: AccountId,
+    ) -> Result<AccountVaultInfo, RpcError>;
+
+    /// Fetches transactions records for specific accounts within a block range.
+    /// Using the `/SyncTransactions` RPC endpoint.
+    ///
+    /// - `block_from`: The starting block number for the range.
+    /// - `block_to`: The ending block number for the range.
+    /// - `account_ids`: The account IDs for which to fetch storage map updates.
+    async fn sync_transactions(
+        &self,
+        block_from: BlockNumber,
+        block_to: Option<BlockNumber>,
+        account_ids: Vec<AccountId>,
+    ) -> Result<TransactionsInfo, RpcError>;
+
+    /// Fetches the network ID of the node.
+    /// Errors:
+    /// - [`RpcError::ExpectedDataMissing`] if the note with the specified root is not found.
     async fn get_network_id(&self) -> Result<NetworkId, RpcError>;
 }
 
@@ -322,6 +368,9 @@ pub enum NodeRpcClientEndpoint {
     SubmitProvenTx,
     SyncNotes,
     GetNoteScriptByRoot,
+    SyncStorageMaps,
+    SyncAccountVault,
+    SyncTransactions,
 }
 
 impl fmt::Display for NodeRpcClientEndpoint {
@@ -343,6 +392,9 @@ impl fmt::Display for NodeRpcClientEndpoint {
             NodeRpcClientEndpoint::SubmitProvenTx => write!(f, "submit_proven_transaction"),
             NodeRpcClientEndpoint::SyncNotes => write!(f, "sync_notes"),
             NodeRpcClientEndpoint::GetNoteScriptByRoot => write!(f, "get_note_script_by_root"),
+            NodeRpcClientEndpoint::SyncStorageMaps => write!(f, "sync_storage_maps"),
+            NodeRpcClientEndpoint::SyncAccountVault => write!(f, "sync_account_vault"),
+            NodeRpcClientEndpoint::SyncTransactions => write!(f, "sync_transactions"),
         }
     }
 }