+++ crates/sqlite-store/src/lib.rs
@@ -55,6 +55,7 @@ use crate::merkle_store::{insert_asset_nodes, insert_storage_map_nodes};
 use crate::sql_error::SqlResultExt;
 
 mod account;
+mod builder;
 mod chain_data;
 mod db_management;
 mod merkle_store;
@@ -62,6 +63,8 @@ mod note;
 mod sql_error;
 mod sync;
 mod transaction;
+
+pub use builder::ClientBuilderSqliteExt;
 
 // SQLITE STORE
 // ================================================================================================
@@ -81,7 +84,7 @@ impl SqliteStore {
 
     /// Returns a new instance of [Store] instantiated with the specified configuration options.
     pub async fn new(database_filepath: PathBuf) -> Result<Self, StoreError> {
-        let sqlite_pool_manager = SqlitePoolManager::new(database_filepath.clone());
+        let sqlite_pool_manager = SqlitePoolManager::new(database_filepath);
         let pool = Pool::builder(sqlite_pool_manager)
             .build()
             .map_err(|e| StoreError::DatabaseError(e.to_string()))?;